SDKS-4586: Refactor Journey and DaVinci plugin modules
This commit centralizes core components by moving them from the `davinci-plugin` and `journey-plugin` modules into the `foundation/orchestrate` module.

The following changes have been made:
- Moved `ContinueNodeAware` and `RequestInterceptor` interfaces into the `orchestrate` module.
- Consolidated `DaVinci` and `Journey` type aliases for `Workflow` into their respective plugin modules, removing the separate files.
- Relocated extension properties `callbacks` and `collectors` for `ContinueNode` into their relevant plugin modules.
