+++ davinci/src/main/kotlin/com/pingidentity/davinci/collector/Collectors.kt
@@ -8,10 +8,10 @@
 package com.pingidentity.davinci.collector
 
 import com.pingidentity.davinci.plugin.Collectors
-import com.pingidentity.davinci.plugin.RequestInterceptor
 import com.pingidentity.davinci.plugin.Submittable
 import com.pingidentity.orchestrate.FlowContext
 import com.pingidentity.orchestrate.Request
+import com.pingidentity.orchestrate.RequestInterceptor
 import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonPrimitive
