--- foundation/davinci-plugin/src/main/kotlin/com/pingidentity/davinci/plugin/DaVinci.kt
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2024 - 2025 Ping Identity Corporation. All rights reserved.
- *
- * This software may be modified and distributed under the terms
- * of the MIT license. See the LICENSE file for details.
- */
-
-
-package com.pingidentity.davinci.plugin
-
-import com.pingidentity.orchestrate.Workflow
-
-typealias DaVinci = Workflow
