+++ foundation/journey-plugin/src/main/kotlin/com/pingidentity/journey/plugin/ContinueNode.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2025 Ping Identity Corporation. All rights reserved.
+ *
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+package com.pingidentity.journey.plugin
+
+import com.pingidentity.orchestrate.ContinueNode
+
+/**
+ * Extension property for Connector class to get a list of collectors.
+ *
+ * @return A list of Collector instances.
+ */
+val ContinueNode.callbacks: List<Callback>
+    get() = this.actions.filterIsInstance<Callback>()
