--- foundation/journey-plugin/src/main/kotlin/com/pingidentity/journey/plugin/Journey.kt
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2024 - 2025 Ping Identity Corporation. All rights reserved.
- *
- * This software may be modified and distributed under the terms
- * of the MIT license. See the LICENSE file for details.
- */
-
-package com.pingidentity.journey.plugin
-
-import com.pingidentity.orchestrate.Workflow
-
-typealias Journey = Workflow
